UserMethodArgumentResolver moved to WebConfig class (validation still doesn't work)
