+++ src/main/java/ru/netology/_72_rest/configuration/WebConfig.java
@@ -1,9 +1,13 @@
 package ru.netology._72_rest.configuration;
 
 import org.springframework.context.annotation.Configuration;
+import org.springframework.core.MethodParameter;
+import org.springframework.web.bind.support.WebDataBinderFactory;
+import org.springframework.web.context.request.NativeWebRequest;
 import org.springframework.web.method.support.HandlerMethodArgumentResolver;
+import org.springframework.web.method.support.ModelAndViewContainer;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-import ru.netology._72_rest.domain.resolver.UserMethodArgumentResolver;
+import ru.netology._72_rest.domain.User;
 
 import java.util.List;
 
@@ -12,5 +16,20 @@ public class WebConfig implements WebMvcConfigurer {
     @Override
     public void addArgumentResolvers(List<HandlerMethodArgumentResolver> argumentResolvers) {
         argumentResolvers.add(new UserMethodArgumentResolver());
+    }
+
+    private static class UserMethodArgumentResolver implements HandlerMethodArgumentResolver {
+        @Override
+        public boolean supportsParameter(MethodParameter parameter) {
+            return parameter.getParameterType().equals(User.class);
+        }
+
+        @Override
+        public Object resolveArgument(MethodParameter parameter, ModelAndViewContainer mavContainer,
+                                      NativeWebRequest webRequest, WebDataBinderFactory binderFactory) {
+            final String name = webRequest.getParameter("user");
+            final String password = webRequest.getParameter("password");
+            return new User(name, password, null);
+        }
     }
 }
