--- src/main/java/ru/netology/_72_rest/domain/resolver/UserMethodArgumentResolver.java
@@ -1,23 +1,0 @@
-package ru.netology._72_rest.domain.resolver;
-
-import org.springframework.core.MethodParameter;
-import org.springframework.web.bind.support.WebDataBinderFactory;
-import org.springframework.web.context.request.NativeWebRequest;
-import org.springframework.web.method.support.HandlerMethodArgumentResolver;
-import org.springframework.web.method.support.ModelAndViewContainer;
-import ru.netology._72_rest.domain.User;
-
-public class UserMethodArgumentResolver implements HandlerMethodArgumentResolver {
-    @Override
-    public boolean supportsParameter(MethodParameter parameter) {
-        return parameter.getParameterType().equals(User.class);
-    }
-
-    @Override
-    public Object resolveArgument(MethodParameter parameter, ModelAndViewContainer mavContainer,
-                                  NativeWebRequest webRequest, WebDataBinderFactory binderFactory) {
-        final String name = webRequest.getParameter("user");
-        final String password = webRequest.getParameter("password");
-        return new User(name, password, null);
-    }
-}
